Add files via upload
Write a function that accepts 1 arg. Compare the passed arg to true using the == operator. If the values are equal, return a message saying that particular arg is equal to true. If the values are not equal, return a message saying that particular arg is not equal to true. Call the function multiple times testing these args: 1, 1.0, "1", and "1.0".