Add files via upload
Write a function that accepts 1 arg. Compare the passed arg to false using the == operator. If the values are equal, return a message saying that particular arg is equal to false. If the values are not equal, return a message saying that particular arg is not equal to false. Call the function multiple times testing these args: 0, 0.0, "0", "0.0", and "".